+++ src/proj2d/mesh/Mesh2dRenderer.ts
@@ -1,6 +1,6 @@
 namespace pixi_projection {
 	let shaderVert =
-		`precision highp float;
+		`precision mediump float;
 attribute vec2 aVertexPosition;
 attribute vec2 aTextureCoord;
